add front pages

+++ src/main/java/com/example/movies/repositories/ConfirmationTokenRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface ConfirmationTokenRepository extends CrudRepository<ConfirmationToken, Long> {
+public interface ConfirmationTokenRepository extends CrudRepository<ConfirmationToken, Integer> {
     Optional<ConfirmationToken> findByToken(String token);
 }
